+++ CHANGELOG.md
@@ -6,6 +6,10 @@ Version 0.1.1
 
 To be released.
 
+### Library
+
+ -  Added `Checkmate.Renderer` module.
+
 
 Version 0.1.0
 -------------